fix: only on windows
