+++ env.go
@@ -8,7 +8,6 @@ import (
 	"net/url"
 	"os"
 	"reflect"
-	"runtime"
 	"strconv"
 	"strings"
 	"time"
@@ -118,7 +117,7 @@ func configure(opts []Options) []Options {
 	// Created options with defaults.
 	opt := Options{
 		TagName:     "env",
-		Environment: toMap(os.Environ(), runtime.GOOS == "windows"),
+		Environment: toMap(os.Environ()),
 		configured:  true,
 	}
 
@@ -134,31 +133,6 @@ func configure(opts []Options) []Options {
 	}
 
 	return []Options{opt}
-}
-
-func toMap(env []string, isWindows bool) map[string]string {
-	r := map[string]string{}
-	for _, e := range env {
-
-		p := strings.SplitN(e, "=", 2)
-
-		if isWindows {
-			// On Windows, environment variables can start with '='. If so, Split at next character.
-			// See env_windows.go in the Go source: https://github.com/golang/go/blob/master/src/syscall/env_windows.go#L58
-			prefixEqualSign := false
-			if len(e) > 0 && e[0] == '=' {
-				e = e[1:]
-				prefixEqualSign = true
-			}
-			p = strings.SplitN(e, "=", 2)
-			if prefixEqualSign {
-				p[0] = "=" + p[0]
-			}
-		}
-
-		r[p[0]] = p[1]
-	}
-	return r
 }
 
 // getTagName returns the tag name.
